Fix bug where a last_post_timestamp of 0 would reset challenge progress
This is only really a problem if you're trying to manually mangle the
database and manually set the next challenge. The code will check
last_post_timestamp to ensure a week has passed since the last time
a challenge was posted. Someone who was trying to force the bot to
post again may thus set last_post_timestamp for a room to 0. However,
the code had a check for if last_post_timestamp was falsely, it
meant we had never posted, and thus to send the first challenge post
in that room.

This commit changes the code to check for None, for 'never posted',
instead of a falsely value, such as 0.